planet: remove socket-activation support.
This change drops systemd socket-activation support.
